chandler/development

+++ backend/dbms.py
@@ -15,13 +15,10 @@ class Database:
     def __init__(self, database_fname: str):
         self.database_fname = database_fname
         self.connection = sqlite3.connect(self.database_fname)
+        self.connection.row_factory = sqlite3.Row
         print("Connection to database established...")
         self.cursor = self.connection.cursor()
         print("Cursor created...")
-
-        # add default payment types
-        self.add_payment_type("Credit Card")
-        self.add_payment_type("Paypal")
 
         self.default_order_status = "Order In Progress"
 
@@ -273,7 +270,7 @@ class Database:
 
     def retrieve_all_product_details_With_Thumbnail_With_Analytics(self):
         """Gets all product details from the db"""
-        self.cursor.execute("SELECT Product.*, ProductThumbnail.ImageName, sum(ProductsInOrder.Quantity), sum(ProductsInOrder.Quantity * ProductsInOrder.PriceSold) FROM Product INNER JOIN ProductThumbnail ON Product.Product_ID = ProductThumbnail.Product_ID LEFT JOIN ProductsInOrder ON Product.Product_ID = ProductsInOrder.Product_ID GROUP BY Product.Product_ID ORDER BY (CURRENT_DATE - Product.DateCreated) ASC")
+        self.cursor.execute("SELECT Product.*, ProductThumbnail.ImageName, sum(ProductsInOrder.Quantity) AS QuantitySold, sum(ProductsInOrder.Quantity * ProductsInOrder.PriceSold) AS TotalEquity FROM Product INNER JOIN ProductThumbnail ON Product.Product_ID = ProductThumbnail.Product_ID LEFT JOIN ProductsInOrder ON Product.Product_ID = ProductsInOrder.Product_ID GROUP BY Product.Product_ID ORDER BY (CURRENT_DATE - Product.DateCreated) ASC")
         return self.cursor.fetchall()
 
     def retrieve_specific_product_details(self, product_id: int):
@@ -303,7 +300,11 @@ class Database:
 
     def retrieve_Top_10_product_details(self):
         """Gets top 10 product details from the db"""
-        self.cursor.execute("SELECT Product.product_id, Product.title, ProductThumbnail.ImageName, sum(ProductsInOrder.Quantity), sum(ProductsInOrder.pricesold * ProductsInOrder.Quantity) FROM ProductsInOrder INNER JOIN Product on Product.product_id = ProductsInOrder.product_id LEFT JOIN ProductThumbnail on ProductThumbnail.product_id = ProductsInOrder.product_id GROUP BY Product.product_id ORDER BY sum(ProductsInOrder.Quantity) DESC LIMIT 10")
+        self.cursor.execute("SELECT Product.Product_ID AS Product_ID,"+
+                            " Product.Title AS Title,"+
+                            " ProductThumbnail.ImageName AS ImageName,"+
+                            " sum(ProductsInOrder.Quantity) AS QuantitySold,"+
+                            " sum(ProductsInOrder.pricesold * ProductsInOrder.Quantity) AS TotalEquity FROM ProductsInOrder INNER JOIN Product on Product.product_id = ProductsInOrder.product_id LEFT JOIN ProductThumbnail on ProductThumbnail.product_id = ProductsInOrder.product_id GROUP BY Product.product_id ORDER BY sum(ProductsInOrder.Quantity) DESC LIMIT 10")
         return self.cursor.fetchall()
 
     def add_product_category(self, category_name: str) -> bool:
@@ -441,28 +442,91 @@ class Database:
         self.cursor.execute("SELECT Product_ID FROM Wishlist WHERE Customer_ID = ?", (customer_id,))
         return self.cursor.fetchall()
 
-    def add_product_to_cart(self, customer_id: int, product_id: int):
+    def add_product_to_cart(self, customer_id: int, product_id: int, quantity: int) -> bool:
         customer_exists = self._does_customer_id_exist(customer_id)
         product_exists = self._does_product_exist(product_id)
         if customer_exists and product_exists:
-            self.cursor.execute("INSERT INTO Cart (Customer_ID, Product_ID) VALUES (?, ?)", (customer_id, product_id))
+            if self._does_cart_product_exist(customer_id, product_id):
+                # If the product is already in the cart, update the quantity
+                self.cursor.execute("UPDATE Cart SET Quantity = Quantity + ? WHERE Customer_ID = ? AND Product_ID = ?", (quantity, customer_id, product_id))
+            else:
+                self.cursor.execute("INSERT INTO Cart (Customer_ID, Product_ID, Quantity) VALUES (?, ?, ?)", (customer_id, product_id, quantity))
             self.connection.commit()
             return True
         return False
 
-    def remove_product_from_cart(self, customer_id: int, product_id: int):
+    def remove_product_from_cart(self, customer_id: int, product_id: int, quantity: int) -> bool:
         """Removes a product from the customer's cart"""
-        self.cursor.execute("DELETE FROM Cart WHERE Customer_ID = ? AND Product_ID = ?", (customer_id, product_id))
-        self.connection.commit()
-        return True
+        customer_exists = self._does_customer_id_exist(customer_id)
+        product_exists = self._does_product_exist(product_id)
+        if customer_exists and product_exists:
+            product_info = self._does_cart_product_exist(customer_id, product_id)
+            if product_info is not None:
+                if product_info[0]['Quantity'] > quantity:
+                    # If the quantity is greater than the quantity to remove, update the quantity
+                    self.cursor.execute("UPDATE Cart SET Quantity = Quantity - ? WHERE Customer_ID = ? AND Product_ID = ?", (quantity, customer_id, product_id))
+                else:
+                    self.cursor.execute("DELETE FROM Cart WHERE Customer_ID = ? AND Product_ID = ?", (customer_id, product_id))
+                self.connection.commit()
+                return True
+        return False
 
-    def get_all_product_ids_in_cart(self, customer_id: int):
+    def get_all_products_in_cart(self, customer_id: int):
         """
         Returns the product ids of all products in the cart. 
         search_product_by_id() can then be used to get the details.
         """
-        self.cursor.execute("SELECT Product_ID FROM Cart WHERE Customer_ID = ?", (customer_id,))
-        return self.cursor.fetchall()
+        self.cursor.execute("SELECT Product.*, ProductThumbnail.ImageName, Cart.Quantity FROM Cart INNER JOIN Product ON Cart.Product_ID = Product.Product_ID INNER JOIN ProductThumbnail ON Product.Product_ID = ProductThumbnail.Product_ID WHERE Customer_ID = ?", (customer_id,))
+        rows = self.cursor.fetchall()
+        return rows
+
+    def get_Specific_Customer_Review(self, customer_id: int, product_id: int):
+        """Gets specific customer review from db"""
+        self.cursor.execute("SELECT ProductReviews.Product_ID AS Product_ID," +
+                                " ProductReviews.Customer_ID AS Customer_ID," +
+                                " ProductReviews.Rating AS Rating," +
+                                " ProductReviews.Review AS Review," +
+                                " ProductReviews.DateOfReview AS DateOfReview" +
+                                ", CustomerUser.Username FROM ProductReviews INNER JOIN CustomerUser ON ProductReviews.Customer_ID = CustomerUser.Customer_ID WHERE ProductReviews.Customer_ID = ? AND ProductReviews.Product_ID = ?", (customer_id, product_id))
+        return self.cursor.fetchone()
+
+    def get_product_review_average(self, product_id: int):
+        """Gets all reviews of a product from db"""
+        self.cursor.execute("SELECT AVG(Rating) AS AverageRating FROM ProductReviews WHERE Product_ID = ? ORDER BY DateOfReview DESC", (product_id,))
+        data = self.cursor.fetchone()
+        return data['AverageRating']
+    
+    def get_all_reviews_of_product_except_customer(self, customer_id: int, product_id: int):
+        """Gets all reviews of a product from db except the customer's review"""
+        if self._does_customer_id_exist(customer_id) is False:
+            self.cursor.execute("SELECT ProductReviews.Product_ID AS Product_ID," +
+                                " ProductReviews.Customer_ID AS Customer_ID," +
+                                " ProductReviews.Rating AS Rating," +
+                                " ProductReviews.Review AS Review," +
+                                " ProductReviews.DateOfReview AS DateOfReview" +
+                                ", CustomerUser.Username AS Username FROM ProductReviews INNER JOIN CustomerUser ON ProductReviews.Customer_ID = CustomerUser.Customer_ID WHERE ProductReviews.Product_ID = ? ORDER BY ProductReviews.DateOfReview ASC", (product_id,))
+        else:
+            self.cursor.execute("SELECT ProductReviews.Product_ID AS Product_ID," +
+                                " ProductReviews.Customer_ID AS Customer_ID," +
+                                " ProductReviews.Rating AS Rating," +
+                                " ProductReviews.Review AS Review," +
+                                " ProductReviews.DateOfReview AS DateOfReview" +
+                                ", CustomerUser.Username FROM ProductReviews INNER JOIN CustomerUser ON ProductReviews.Customer_ID = CustomerUser.Customer_ID WHERE ProductReviews.Product_ID = ? AND ProductReviews.Customer_ID != ? ORDER BY ProductReviews.DateOfReview ASC", (product_id, customer_id,))
+        data = self.cursor.fetchall()
+        if len(data) == 0:
+            return None
+        return data
+        
+    def add_review_to_product(self, customer_id: int, product_id: int, rating:int, review: str) -> bool:
+        """Inserts a review for a product into the db"""
+        customer_exists = self._does_customer_id_exist(customer_id)
+        product_exists = self._does_product_exist(product_id)
+        review_exists = self._does_Review_Of_Product_exist(customer_id, product_id)
+        if customer_exists and product_exists and not review_exists:
+            self.cursor.execute("INSERT INTO ProductReviews (Customer_ID, Product_ID, Rating, Review, DateOfReview) VALUES (?, ?, ?, ?, strftime('%Y-%m-%d', 'now'))", (customer_id, product_id, rating, review))
+            self.connection.commit()
+            return True
+        return False
 
     def add_payment_type(self, payment_type_name: str):
         """Adds a new payment type to the db"""
@@ -718,6 +782,21 @@ class Database:
         self.cursor.execute("SELECT Order_ID FROM Orders WHERE Customer_ID = ?", (customer_id,))
         return self.cursor.fetchall()
 
+    def get_all_orders(self, op_status_filter: str = None, op_orderId_filter: int = None):
+        """Gets all orders from the db. If op_status_filter is provided, only orders with that status are returned.
+        If op_orderId_filter is provided, only orders with that order id are returned. if both are provided, 
+        only orders with that status and order id are returned."""
+        
+        if op_status_filter is not None and op_orderId_filter is not None:
+            self.cursor.execute("SELECT * FROM Orders WHERE StatusName = ? AND Order_ID = ?", (op_status_filter, op_orderId_filter))
+        elif op_status_filter is not None:
+            self.cursor.execute("SELECT * FROM Orders WHERE StatusName = ?", (op_status_filter,))
+        elif op_orderId_filter is not None:
+            self.cursor.execute("SELECT * FROM Orders WHERE Order_ID = ?", (op_orderId_filter,))
+        else:
+            self.cursor.execute("SELECT * FROM Orders")
+        return self.cursor.fetchall()
+
     def cancel_order(self, order_id: int):
         self.cursor.execute("UPDATE Orders SET StatusName = ? WHERE Order_ID = ?", ("Cancelled", order_id))
         self.connection.commit()
@@ -736,6 +815,11 @@ class Database:
     def get_order_status(self, order_id: int):
         self.cursor.execute("SELECT StatusName FROM Orders WHERE Order_ID = ?", (order_id,))
         return self.cursor.fetchone()[0]
+    
+    def get_all_order_statuses(self):
+        """Gets all order statuses from the db"""
+        self.cursor.execute("SELECT * FROM OrderStatus")
+        return self.cursor.fetchall()
 
     def _hash_new_password(self, password: str) -> tuple[str, str]:
         """Hashes salted password w/ bcrypt"""
@@ -783,6 +867,39 @@ class Database:
         if len(self.cursor.fetchall()) == 0:
             return False
         return True
+
+    def _does_cart_product_exist(self, customer_id: int, product_id: int):
+        """Checks if product is in the cart for the customer"""
+        self.cursor.execute("SELECT Customer_ID, Product_ID Quantity FROM Cart WHERE Customer_ID = ? AND Product_ID = ?", (customer_id, product_id,))
+        data = self.cursor.fetchall()
+        if len(data) != 0:
+            return data
+        return None
+    
+    def _does_Wishlist_Product_exist(self, customer_id: int, product_id: int):
+        """Checks if product is in the cart for the customer"""
+        self.cursor.execute("SELECT Customer_ID, Product_ID FROM Wishlist WHERE Customer_ID = ? AND Product_ID = ?", (customer_id, product_id,))
+        data = self.cursor.fetchall()
+        if len(data) != 0:
+            return data
+        return None
+
+    def _does_Review_Of_Product_exist(self, customer_id: int, product_id: int) -> bool:
+        """Checks if product is in the cart for the customer"""
+        self.cursor.execute("SELECT Customer_ID, Product_ID FROM ProductReviews WHERE Customer_ID = ? AND Product_ID = ?", (customer_id, product_id,))
+        data = self.cursor.fetchall()
+        if len(data) != 0:
+            return True
+        return False
+    
+    def _does_Product_Exist_In_Customer_Orders(self, customer_id: int, product_id: int) -> bool:
+        """Checks if product has been ordered by the customer"""
+        self.cursor.execute("SELECT Orders.Order_ID, Orders.Customer_ID, ProductsInOrder.Product_ID FROM Orders INNER JOIN ProductsInOrder ON Orders.Order_ID = ProductsInOrder.Order_ID WHERE Orders.Customer_ID = ? AND ProductsInOrder.Product_ID = ?", (customer_id, product_id,))
+        data = self.cursor.fetchall()
+        if len(data) != 0:
+            return True
+        return False
+        
 
     def _new_customer_id(self) -> int:
         """generate new customer id artifical key"""
